backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.12
httpx==0.27.2
sse-starlette==2.1.3
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic==2.9.2
aiofiles==24.1.0
structlog==24.4.0
rich==13.8.1

# RAG dependencies
chromadb==0.5.7
sentence-transformers==3.1.1
pypdf==4.3.1
python-docx==1.1.2
tiktoken==0.7.0

# LangGraph (optional orchestration layer)
langgraph==0.2.28
langchain==0.3.0
langchain-community==0.3.0
langchain-openai==0.2.0
